Cookies: move cookie parsing into its own parser file
