+++ lib/src/parser/http11.rs
@@ -4,6 +4,7 @@
 use sozu_command::buffer::Buffer;
 use network::buffer_queue::BufferQueue;
 use network::protocol::StickySession;
+use parser::cookies::{CookieValue, parse_request_cookies};
 
 use nom::{HexDisplay,IResult,Offset};
 use nom::IResult::*;
@@ -237,53 +238,6 @@ fn is_hostname_char(i: u8) -> bool {
 }
 
 named!(pub hostname_and_port<(&[u8],Option<&[u8]>)>, pair!(take_while!(is_hostname_char), opt!(complete!(preceded!(tag!(":"), digit)))));
-
-#[derive(Debug)]
-pub struct CookieValue<'a> {
-  name: &'a [u8],
-  value: &'a [u8],
-  semicolon: Option<&'a [u8]>,
-  spaces: &'a [u8]
-}
-
-impl<'a> CookieValue<'a> {
-  pub fn get_full_length(&self) -> usize {
-    let semicolon = if self.semicolon.is_some() { 1 } else { 0 };
-    let space = self.spaces.len();
-
-    self.name.len() + self.value.len() + semicolon + space + 1 // +1 is for =
-  }
-}
-
-pub fn is_cookie_value_char(chr: u8) -> bool {
-  // chars: (space) , ;
-  chr != 32 && chr != 44 && chr != 59
-}
-
-named!(pub single_request_cookie<CookieValue>,
-  do_parse!(
-    name: take_until_and_consume!("=") >>
-    value: take_while!(is_cookie_value_char) >>
-    semicolon: opt!(complete!(tag!(";"))) >>
-    spaces: complete!(take_while!(is_space)) >>
-    (CookieValue {
-      name: name,
-      value: value,
-      semicolon: semicolon,
-      spaces: spaces
-    })
-  )
-);
-
-pub fn request_cookies(input: &[u8]) -> Option<Vec<CookieValue>> {
-  let res: IResult<&[u8], Vec<CookieValue>> = many0!(input, single_request_cookie);
-
-  if let IResult::Done(_, o) = res {
-    Some(o)
-  } else {
-    None
-  }
-}
 
 use std::str::{from_utf8, FromStr};
 use nom::{Err,ErrorKind,Needed};
@@ -527,7 +481,7 @@ impl<'a> Header<'a> {
         }
       },
       b"cookie" => {
-        match request_cookies(self.value) {
+        match parse_request_cookies(self.value) {
           Some(cookies) => HeaderValue::Cookie(cookies),
           None          => HeaderValue::Error
         }
@@ -603,7 +557,7 @@ impl<'a> Header<'a> {
   }
 
   pub fn remove_sticky_cookie_in_request(&self, buf: &[u8], offset: usize) -> Vec<BufferMove> {
-    if let Some(cookies) = request_cookies(self.value) {
+    if let Some(cookies) = parse_request_cookies(self.value) {
       // if we don't find the cookie, don't go further
       if let Some(sozu_balance_position) = cookies.iter().position(|cookie| &cookie.name[..] == b"SOZUBALANCEID") {
         // If we have only one cookie and that's the one, then we drop the whole header
@@ -3062,35 +3016,6 @@ mod tests {
     };
 
     assert!(header.must_mutate());
-  }
-
-  #[test]
-  fn header_cookies_request_cookies() {
-    let none_cookies = request_cookies(b"FOOBAR").unwrap();
-    let some_cookies = request_cookies(b"FOO=BAR;BAR=FOO; SOZUBALANCEID=0;   SOZU=SOZU").unwrap();
-
-    assert_eq!(none_cookies.len(), 0);
-    assert_eq!(some_cookies.len(), 4);
-
-    assert_eq!(some_cookies[0].name, b"FOO");
-    assert_eq!(some_cookies[0].value, b"BAR");
-    assert_eq!(some_cookies[0].semicolon.is_some(), true);
-    assert_eq!(some_cookies[0].spaces.len(), 0);
-
-    assert_eq!(some_cookies[1].name, b"BAR");
-    assert_eq!(some_cookies[1].value, b"FOO");
-    assert_eq!(some_cookies[1].semicolon.is_some(), true);
-    assert_eq!(some_cookies[1].spaces.len(), 1);
-
-    assert_eq!(some_cookies[2].name, b"SOZUBALANCEID");
-    assert_eq!(some_cookies[2].value, b"0");
-    assert_eq!(some_cookies[2].semicolon.is_some(), true);
-    assert_eq!(some_cookies[2].spaces.len(), 3);
-
-    assert_eq!(some_cookies[3].name, b"SOZU");
-    assert_eq!(some_cookies[3].value, b"SOZU");
-    assert_eq!(some_cookies[3].semicolon.is_some(), false);
-    assert_eq!(some_cookies[3].spaces.len(), 0);
   }
 
   #[test]
