+++ lib/src/parser/mod.rs
@@ -1,1 +1,2 @@
 pub mod http11;
+pub mod cookies;
